[fixed] tc fixes
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1226 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/dots/DamageOverTimeMap.cpp
@@ -15,7 +15,7 @@ void DamageOverTimeMap::addDot(CreatureObject* attacker, CreatureObject* victim,
 	if (defense > 0)
 		dotReductionMod -= (float) defense / 125.0f;
 
-	int redStrength = strength * dotReductionMod;
+	int redStrength = (int)(strength * dotReductionMod);
 	float redPotency = potency * dotReductionMod;
 
 	bool dotAdded =  dotList->addDot(victim, dotID, duration, dotType, pool, redStrength,redPotency);
